requirements.txt
pytorch_lightning==2.0.0
torch==2.0.0
scikit-learn==1.2.2
hydra-core==1.3.2
omegaconf==2.3.0

scipy==1.10.1
pytube==15.0.0
moviepy==1.0.3